refactor: simplify cli and benchmark tooling
