+++ CONTRIBUTING.rst
@@ -186,7 +186,7 @@ A reminder for the maintainers on how to deploy.
 Make sure all your changes are committed (including an entry in HISTORY.rst).
 Then run::
 
-$ bumpversion patch # possible: major / minor / patch
+Update the version in pyproject.toml and json2xml/__init__.py
 $ git push
 $ git push --tags
