+++ setup.cfg
@@ -1,29 +1,6 @@
-[bumpversion]
-current_version = 6.1.0
-commit = True
-tag = True
-
-[bumpversion:file:setup.py]
-search = version='{current_version}'
-replace = version='{new_version}'
-
-[bumpversion:file:json2xml/__init__.py]
-search = __version__ = '{current_version}'
-replace = __version__ = '{new_version}'
-
-[bdist_wheel]
-universal = false
-
 [flake8]
 exclude = docs
 max-line-length=120
-
-[aliases]
-# Define setup.py command aliases here
-
-[tool.ty]
-# Ty configuration is now in pyproject.toml
-
 
 [coverage:run]
 relative_files = True
